refactor AUTOSolution to its own module
